backend/requirements.txt
flask
flask-cors
pdfminer.six
langchain
langchain-community
langchain-huggingface
langchain-ollama
faiss-cpu
huggingface_hub
sentence-transformers
pdfplumber
pandas
numpy
pypdf